% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_data_quality}
\alias{check_data_quality}
\title{check_data_quality}
\usage{
check_data_quality(n_obs, dilution, id_columns)
}
\arguments{
\item{n_obs}{number of observations}

\item{dilution}{numeric array of log10 well dilution factors
(0 = no dilution, -1 = 10-fold dilution, etc).}

\item{id_columns}{list of columns in the data containing
ids for each observation (all of which must be integer vectors
of length equal to the number of observations)}
}
\value{
TRUE
}
\description{
Verify that data does not have obvious errors
before passing to Stan fitting functions
}
