% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{format_est_ci}
\alias{format_est_ci}
\title{format_est_ci}
\usage{
format_est_ci(estimate, lower_ci, upper_ci, significant_digits = 3)
}
\arguments{
\item{estimate}{the central estimate}

\item{lower_ci}{the lower bound of the CI}

\item{upper_ci}{the upper bound of the CI}

\item{significant_digits}{how many digits to treat as
significant in rounding / truncating}
}
\description{
format a central estimate (e.g. a mode)
and its bounding interval (CI, e.g. a
credible interval or a confidence
interval) for in-text display
as:
}
\details{
central \link{lower CI, upper CI}
}
