% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{get_random_draws}
\alias{get_random_draws}
\title{get_random_draws}
\usage{
get_random_draws(
  mcmc_draws,
  n_samples,
  draw_id_column = ".draw",
  replace = FALSE
)
}
\arguments{
\item{mcmc_draws}{mcmc chains in spread_draws tidy
tibble format from tidybayes}

\item{n_samples}{how many draws to sample}

\item{draw_id_column}{name of the draw id column,
default ".draws" (as is default in tidybayes)}

\item{replace}{whether to sample with replacement.
Default FALSE.}
}
\value{
tibble of subsampled draws
}
\description{
Given a tidy dataframe of
MCMC results, get a random
subsample of draws, returned
as another tidy dataframe
with the same scheme but fewer
rows
}
\details{
This is important to have since the same draw
will occur on multiple lines of of the data frame
}
