% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{pango_to_who}
\alias{pango_to_who}
\title{pango_to_who}
\usage{
pango_to_who(pango_lineage, pango_if_fail = FALSE)
}
\arguments{
\item{pango_lineage}{PANGO lineage(s) of the variant(s)}

\item{pango_if_fail}{Flag for what to do if lookup fails. If TRUE
(default) return the PANGO lineage inputed. If FALSE,
return NA.}

\item{retain_pango}{flag for whether to retain the
PANGO lineage in parentheses after to WHO name. Default FALSE.}
}
\value{
vector of WHO variant names, possibly with
PANGO lineage in parentheses
}
\description{
converts a vector of PANGO lineage designations
to a vector of WHO Greek letter variant names
}
