% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{plotting_params}
\alias{plotting_params}
\title{plotting_params}
\usage{
plotting_params(param_to_get)
}
\arguments{
\item{param_to_get}{which parameter is desired}
}
\value{
parameter value
}
\description{
plotting_params
}
