% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{scale_color_variant}
\alias{scale_color_variant}
\title{scale_color_variant}
\usage{
scale_color_variant(...)
}
\arguments{
\item{...}{other parameters passed to scale_color_manual()}
}
\value{
color scale
}
\description{
default color scale for variants
for this project
}
