% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{scale_fill_variant}
\alias{scale_fill_variant}
\title{scale_fill_variant}
\usage{
scale_fill_variant(...)
}
\arguments{
\item{...}{other parameters passed to scale_fill_manual()}
}
\value{
fill scale
}
\description{
default fill scale for variants
for this project
}
