% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R, R/plotting_style.R
\name{stat_default_eye}
\alias{stat_default_eye}
\title{stat_default_eye}
\usage{
stat_default_eye(size = 15, interval_size_ratio = 2.5, ...)

stat_default_eye(size = 15, interval_size_ratio = 2.5, ...)
}
\arguments{
\item{size}{size argument for the eyeplot, default 20}

\item{interval_size_ratio}{how much greater should the
linewidth of the smaller (default 68\%) credible interval
be compared to that of the larger (default 95\%) credible
interval? Default 2.5}

\item{...}{other parameters passed to stat_eye()}
}
\value{
list of stats that can be added to a ggplot object

list of stats that can be added to a ggplot object
}
\description{
default eyeplot for figures in this project

default eyeplot for figures in this project
}
