% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{stat_exp_curve}
\alias{stat_exp_curve}
\title{Plot an exponential decay curve using
ggplot}
\usage{
stat_exp_curve(
  mapping = NULL,
  data = NULL,
  geom = "line",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  n = 50,
  xmin = 0,
  xmax = 1,
  base = exp(1),
  ...
)
}
\arguments{
\item{n}{number of distinct points to plot (default 50)}

\item{xmin}{first x-value to plot (default 0)}

\item{xmax}{last x-value to plot (default 1)}

\item{base}{base of the exponential (default e)}
}
\description{
Plot an exponential decay curve using
ggplot
}
