% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{StatExpCurve}
\alias{StatExpCurve}
\title{StatExpCurve}
\format{
An object of class \code{StatExpCurve} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatExpCurve
}
\description{
Stat class for exponential growth/decay
curves in ggplot
}
\keyword{datasets}
