% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\docType{data}
\name{variant_colors}
\alias{variant_colors}
\title{variant_colors}
\format{
An object of class \code{list} of length 7.
}
\usage{
variant_colors
}
\description{
color scheme for variants
of concern
}
\details{
Palette generated with Colorgorical
http://vrl.cs.brown.edu/color
}
\keyword{datasets}
